Former-commit-id: c46602a113bba5f7c626669a024a86b0681ce519 [formerly 0aaf0682234524f019c5b09f9337aa6e6976e820]
Former-commit-id: 094aa97d5e5575d6bb5498884d2209e5981f0a56

+++ soccer_fieldline_detection/include/soccer_fieldline_detection/camera.hpp
@@ -18,8 +18,8 @@ class Camera {
     float max_line_distance = 10;   // Maximum distance to project the current line, if the endpoint is above the horizon
 
     // Laser scan data laser_scan_angle_delta = deg2rad(3);
-    double angmin = 0;
-    double angmax = M_PI;
+    float angmin = 0;
+    float angmax = M_PI;
 
     // Camera Image (type Camera Image)
     // image
